stdrenderer: move to its own module
